Refactored: Cache instrumentation for ActiveSupport.
